+++ src/layer/domain/data/config/scope.ts
@@ -8,23 +8,26 @@ import { extend, overwrite } from 'spica/assign';
 const { match } = router.helpers();
 
 export function scope(
-  config: Config,
+  option: Option,
   path: {
     orig: URL.Pathname<StandardURL>;
     dest: URL.Pathname<StandardURL>;
   }
 ): Maybe<Config> {
-  const scope = { '/': {}, ...config.scope };
-  for (const pattern of Object.keys(scope).reverse()) {
+  const record = { '/': {}, ...option.scope };
+  for (const pattern of Object.keys(record).reverse()) {
+    const opt: Option | undefined = record[pattern];
     switch (+match(pattern, path.orig) + +match(pattern, path.dest)) {
       case 0:
         continue;
       case 1:
-        return Nothing;
+        if (opt === undefined || opt.isolation) return Nothing;
+        continue;
     }
-    const option: Option | undefined = scope[pattern];
-    return option
-      ? Just(new Config(overwrite(extend({}, config, option), { scope: undefined })))
+    return opt
+      ? opt.scope
+        ? scope(overwrite({}, { ...option, scope: undefined }, opt), path)
+        : Just(new Config({ ...extend({}, option, opt), scope: undefined }))
       : Nothing;
   }
   return Nothing;