+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.43.0
 
+- Refine scope option.
+- Add `isolation` option.
 - Extend `fetch.rewrite` option to add cache parameter.
 - Change `X-Pjax` header value to JSON string of `areas` option.